backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.10
djangorestframework==3.16.1
geographiclib==2.1
geopy==2.4.1
idna==3.11
numpy==2.4.1
pandas==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
shapely==2.1.2
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
